fix(EditableText): remove 'align' from non-block element toolbar options
When a non-block element (eg. p tag) gets given an align, it seems to stick for the immediate changes, but then reverts when you the saved value is reinstated. I have a feeling this can be fixed in the app, but I am removing it from the content until then.